refactor(tests): consolidate dummy classes and fixtures in conftest.py
